requirements.txt
torch
tensorboardX
pyyaml
ray
ray[tune]
gym
lz4
hiredis
scipy